Update order and license table column types

+++ config/Data/Migrations/20200206005136_add_orders_table.php
@@ -40,7 +40,7 @@ class AddOrdersTable extends AbstractMigration
             )
             ->addColumn(
                 'stripe_amount',
-                'integer',
+                'float',
                 ['signed' => false]
             )
             ->addColumn(
